requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
asgiref==3.2.10           # via django
authlib==0.14.3           # via -r requirements.in, loginpass
certifi==2020.6.20        # via requests
cffi==1.14.1              # via cryptography
chardet==3.0.4            # via requests
cryptography==3.0         # via authlib
django-appconf==1.0.4     # via django-compressor
django-compressor==2.4    # via -r requirements.in
django-crispy-forms==1.9.2  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-sesame==2.0        # via -r requirements.in
django-tables2==2.3.1     # via -r requirements.in
django==3.1               # via -r requirements.in, django-appconf, django-sesame, django-tables2
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
loginpass==0.5            # via -r requirements.in
mysqlclient==2.0.1        # via -r requirements.in
pycparser==2.20           # via cffi
pytz==2020.1              # via django
rcssmin==1.0.6            # via django-compressor
requests==2.24.0          # via loginpass
rjsmin==1.1.0             # via django-compressor
six==1.15.0               # via cryptography, django-compressor
sqlparse==0.3.1           # via django
urllib3==1.25.10          # via requests
wheel==0.35.1             # via -r requirements.in
whitenoise==5.2.0         # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
